pyproject.toml
[tool.poetry]
name = "ntpu-linebot"
version = "0.1.0"
description = "國立臺北大學的查詢小工具"
authors = ["garyellow <gary20011110@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12"
requests = "^2.31.0"
aiohttp = {extras = ["speedups"], version = "^3.9.3"}
beautifulsoup4 = "^4.12.3"
line-bot-sdk = "^3.9.0"
cachetools = "^5.3.2"
sanic = "^23.12.1"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
setuptools = "^69.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
